Add JUnit runner for checking specs.
This runner makes it possible to use a JUnit test for a single page
test. A spec file is assigned to a test class. Every object spec of
this file is treated as a single test. This automatically provides
JUnit test formatting in IDEs, build tools and build servers. E.g.
you get a test tree in IntelliJ idea that looks like this

    + GalesSpecRunnerIT$ValidSpec
       + first_paragraph
         - height > 10 px
         - width < 1000 px
       + second_paragraph
         - height > 10 px
         - width < 1000 px

This makes it very easy to spot errors.